+++ NEWS.md
@@ -1,6 +1,8 @@
 ## Version 1.1.1
 
 * `heplot.cancor()` gains a `rev.axes` argument, similar to that in `heplot.candisc()`
+* `rev.axes` also added to `heplot3d*()` functions
+* Added `confusion()` to calculate confusion matrices for LDA/QDA
 
 ## Version 1.1.0